Chore/nilcheck (#401)
* Now ExecuteWithContext and fetchMatchingRule are not allow nil KnowledgeBase

* Fixed concurrency test that seems number overflows uint, and added NilKnowledgeBasePanic_test.go

---------

+++ examples/Concurrency_test.go
@@ -66,9 +66,9 @@ var (
 type Vibonaci struct {
 	fmt.Stringer
 	Count int
-	A     uint
-	B     uint
-	C     uint
+	A     uint64
+	B     uint64
+	C     uint64
 }
 
 func startThread(threadName string) {
